FullStack_Projects/Recipe/angularapp/src/app/recipe-list/recipe-list.component.ts
// recipe-list.component.ts
import { Component, OnInit } from '@angular/core';
import { Recipe } from '../models/recipe.model';
import { RecipeService } from '../services/recipe.service';
import { Router } from '@angular/router';


@Component({
  selector: 'app-recipe-list',
  templateUrl: './recipe-list.component.html',
  styleUrls: ['./recipe-list.component.css']
})
export class RecipeListComponent implements OnInit {
  recipes: Recipe[] = [];

  constructor(private recipeService: RecipeService,private router: Router) { }

  ngOnInit(): void {
    this.loadRecipes();
  }

  loadRecipes(): void {
    this.recipeService.getRecipes().subscribe(recipes => this.recipes = recipes);
  }

  Delete(recipeId: number): void {
    // Navigate to confirm delete page with the recipe ID as a parameter
    this.router.navigate(['/confirmDelete', recipeId]);
  }
}

